Move scrypto load logic into ScryptoInterpreter
