+++ radix-engine/src/engine/interpreters/mod.rs
@@ -1,0 +1,4 @@
+mod scrypto_interpreter;
+
+pub use crate::engine::wasm_runtime::*;
+pub use scrypto_interpreter::*;
