+++ radix-engine/src/engine/wasm_runtime.rs
@@ -1,11 +1,10 @@
-use crate::engine::{HeapRENode, SystemApi};
-use crate::engine::{LockFlags, RuntimeError};
+use crate::engine::*;
+use crate::types::*;
 use crate::fee::*;
 use crate::model::{
     Component, ComponentInfoSubstate, ComponentStateSubstate, InvokeError, KeyValueStore,
     RuntimeSubstate,
 };
-use crate::types::*;
 use crate::wasm::*;
 use scrypto::core::FnIdent;
