+++ radix-engine/tests/metering.rs
@@ -153,13 +153,13 @@ fn test_basic_transfer() {
         + 0 /* blobs */
         + 2500 /* create_node */
         + 1698 /* decode_manifest */
-        + 1700 /* drop_lock */
+        + 2300 /* drop_lock */
         + 1000 /* drop_node */
         + 0 /* instantiate_wasm */
         + 2215 /* invoke_function */
-        + 2700 /* lock_substate */
+        + 3600 /* lock_substate */
         + 1100 /* read_owned_nodes */
-        + 5500 /* read_substate */
+        + 10000 /* read_substate */
         + 5200 /* run_function */
         + 338152 /* run_wasm */
         + 566 /* verify_manifest */
@@ -195,7 +195,7 @@ fn test_publish_large_package() {
     receipt.expect_commit_success();
 
     // Assert
-    assert_eq!(4287061, receipt.execution.fee_summary.cost_unit_consumed);
+    assert_eq!(4289061, receipt.execution.fee_summary.cost_unit_consumed);
 }
 
 #[test]
